requirements.txt
Django==6.0.2
django-environ==0.12.0
django-cors-headers==4.7.0
django-mptt==0.18.0
django-filter==25.1
django-unfold
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
drf-spectacular==0.29.0
gunicorn==23.0.0
psycopg2-binary==2.9.10
requests==2.32.3
PyYAML==6.0.3